+++ doc/whats-new.rst
@@ -21,6 +21,11 @@ v2022.07.0 (unreleased)
 
 New Features
 ~~~~~~~~~~~~
+
+- Add :py:meth:`Dataset.set_xindex` and :py:meth:`Dataset.drop_indexes` and
+  their DataArray counterpart for setting and dropping pandas or custom indexes
+  given a set of arbitrary coordinates. (:pull:`6971`)
+  By `Benoît Bovy <https://github.com/benbovy>`_ and `Justus Magin <https://github.com/keewis>`_.
 - Enable taking the mean of dask-backed :py:class:`cftime.datetime` arrays
   (:pull:`6556`, :pull:`6940`).  By `Deepak Cherian
   <https://github.com/dcherian>`_ and `Spencer Clark